* warning fix blaaa
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@4258 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/ascent-world/Player.cpp
@@ -1457,7 +1457,7 @@ void Player::smsg_InitialSpells()
 	PlayerCooldownMap::iterator itr, itr2;
 
 	uint16 spellCount = (uint16)mSpells.size();
-	uint32 itemCount = m_cooldownMap[0].size() + m_cooldownMap[1].size();
+	size_t itemCount = m_cooldownMap[0].size() + m_cooldownMap[1].size();
 	uint32 mstime = getMSTime();
 	size_t pos;
